version2018/municipedia/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
django-tagulous==0.12.0
idna==2.6
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
requests==2.18.4
urllib3==1.22
